Bugs fixed and safer code

+++ src/HOTP.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2014-2018 Spomky-Labs
+ * Copyright (c) 2014-2019 Spomky-Labs
  *
  * This software may be modified and distributed under the terms
  * of the MIT license.  See the LICENSE file for details.
@@ -66,7 +66,7 @@ final class HOTP extends OTP implements HOTPInterface
 
     private function getWindow(?int $window): int
     {
-        return (int) abs($window ?? 0);
+        return abs($window ?? 0);
     }
 
     private function verifyOtpWithWindow(string $otp, int $counter, ?int $window): bool
